+++ builder/vmware/common/step_compact_disk_test.go
@@ -2,8 +2,6 @@ package common
 
 import (
 	"context"
-	"io/ioutil"
-	"os"
 	"testing"
 
 	"github.com/hashicorp/packer/helper/multistep"
@@ -17,25 +15,8 @@ func TestStepCompactDisk(t *testing.T) {
 	state := testState(t)
 	step := new(StepCompactDisk)
 
-	// Create a fake vmdk file for disk file size operations
-	diskFile, err := ioutil.TempFile("", "disk.vmdk")
-	if err != nil {
-		t.Fatalf("Error creating fake vmdk file: %s", err)
-	}
-
-	diskFullPath := diskFile.Name()
-	defer os.Remove(diskFullPath)
-
-	content := []byte("I am the fake vmdk's contents")
-	if _, err := diskFile.Write(content); err != nil {
-		t.Fatalf("Error writing to fake vmdk file: %s", err)
-	}
-	if err := diskFile.Close(); err != nil {
-		t.Fatalf("Error closing fake vmdk file: %s", err)
-	}
-
-	// Set up required state
-	state.Put("disk_full_paths", []string{diskFullPath})
+	diskFullPaths := []string{"foo"}
+	state.Put("disk_full_paths", diskFullPaths)
 
 	driver := state.Get("driver").(*DriverMock)
 
@@ -51,7 +32,7 @@ func TestStepCompactDisk(t *testing.T) {
 	if !driver.CompactDiskCalled {
 		t.Fatal("should've called")
 	}
-	if driver.CompactDiskPath != diskFullPath {
+	if driver.CompactDiskPath != "foo" {
 		t.Fatal("should call with right path")
 	}
 }