+++ common/shell-local/communicator_test.go
@@ -19,12 +19,13 @@ func TestCommunicator(t *testing.T) {
 		return
 	}
 
-	c := &Communicator{}
+	c := &Communicator{
+		ExecuteCommand: []string{"/bin/sh", "-c", "echo foo"},
+	}
 
 	var buf bytes.Buffer
 	cmd := &packer.RemoteCmd{
-		Command: "/bin/echo foo",
-		Stdout:  &buf,
+		Stdout: &buf,
 	}
 
 	if err := c.Start(cmd); err != nil {